mv _transition_index to .util
rename axes to ax
