+++ calour/heatmap/heatmap.py
@@ -17,42 +17,10 @@ import numpy as np
 from ..transforming import log_n
 from ..database import _get_database_class
 from .._dendrogram import plot_tree
+from ..util import _to_list, _transition_index
 
-from ..util import _to_list
 
 logger = getLogger(__name__)
-
-
-def _transition_index(l):
-    '''Return the transition index and current value of the list.
-
-    Examples
-    -------
-    >>> l = ['a', 'a', 'b']
-    >>> list(_transition_index(l))
-    [(2, 'a'), (3, 'b')]
-    >>> l = ['a', 'a', 'b', 1, 2, None, None]
-    >>> list(_transition_index(l))
-    [(2, 'a'), (3, 'b'), (4, 1), (5, 2), (7, None)]
-
-    Parameters
-    ----------
-    l : Iterable of arbitrary objects
-
-    Yields
-    ------
-    tuple of (int, arbitrary)
-        the transition index, the item value
-    '''
-    it = enumerate(l)
-    i, item = next(it)
-    item = str(type(item)), item
-    for i, current in it:
-        current = str(type(current)), current
-        if item != current:
-            yield i, item[1]
-            item = current
-    yield i + 1, item[1]
 
 
 def _create_plot_gui(exp, gui='cli', databases=('dbbact',), tree_size=0):
@@ -109,7 +77,7 @@ def _create_plot_gui(exp, gui='cli', databases=('dbbact',), tree_size=0):
 def heatmap(exp, sample_field=None, feature_field=False, yticklabels_max=100,
             xticklabel_rot=45, xticklabel_len=10, yticklabel_len=15,
             title=None, clim=None, cmap=None,
-            axes=None, rect=None,  transform=log_n,
+            ax=None, rect=None,  transform=log_n,
             show_legend=False, **kwargs):
     '''Plot a heatmap for the experiment.
 
@@ -147,7 +115,7 @@ def heatmap(exp, sample_field=None, feature_field=False, yticklabels_max=100,
         None (default) to use mpl default color map. str to use colormap named str.
     title : None or str (optional)
         None (default) to show experiment description field as title. str to set title to str.
-    axes : matplotlib ``AxesSubplot`` object or None (optional)
+    ax : matplotlib Axes object or None (optional)
         The axes where the heatmap is plotted. None (default) to create a new figure and
         axes to plot heatmap into the axes
     rect : tuple of (int, int, int, int) or None (optional)
@@ -175,10 +143,10 @@ def heatmap(exp, sample_field=None, feature_field=False, yticklabels_max=100,
         logger.debug('transform exp with %r with param %r' % (transform, kwargs))
         data = transform(exp, inplace=False, **kwargs).data
 
-    if axes is None:
+    if ax is None:
         fig, ax = plt.subplots()
     else:
-        fig, ax = axes.get_figure(), axes
+        fig = ax.get_figure()
 
     # step 2. plot heatmap.
     if title is not None:
@@ -189,7 +157,7 @@ def heatmap(exp, sample_field=None, feature_field=False, yticklabels_max=100,
     # plot the heatmap
     image = ax.imshow(data.transpose(), aspect='auto', interpolation='nearest', cmap=cmap, clim=clim)
     if show_legend:
-        _figure_color_bar(exp, image, fig, axes=None, log_scale=(transform == log_n))
+        _imshow_legend(exp, image, fig, ax=None, log_scale=(transform == log_n))
     # set the initial zoom window if supplied
     if rect is not None:
         ax.set_xlim((rect[0], rect[1]))
@@ -270,7 +238,7 @@ def heatmap(exp, sample_field=None, feature_field=False, yticklabels_max=100,
     return fig
 
 
-def _figure_color_bar(exp, image, fig, axes, log_scale=True, num_ticks=8, aspect=100):
+def _imshow_legend(exp, image, fig, ax, log_scale=True, num_ticks=8, aspect=100):
     '''Plot the heatmap colorbar scale
 
     Parameters
@@ -280,8 +248,8 @@ def _figure_color_bar(exp, image, fig, axes, log_scale=True, num_ticks=8, aspect
         We use it to see whether the scale should be relative(normalized) or absolute
     fig : matplotlib.Figure
         The figure to plot the colorbar in
-    axes : matplotlib.Axis
-        The axis to plot into
+    ax : matplotlib Axes
+        The axes to plot into
     log_scale : bool (optional)
         True (default) to plot log scale colorbar
         False to plot linear scale
@@ -291,7 +259,7 @@ def _figure_color_bar(exp, image, fig, axes, log_scale=True, num_ticks=8, aspect
     clim = image.get_clim()
     ticks = np.linspace(clim[0], clim[1], num_ticks)
     ticks = [eval("%.0e" % (x)) for x in ticks]
-    cb = fig.colorbar(image, ax=axes, ticks=ticks, aspect=aspect)
+    cb = fig.colorbar(image, ax=ax, ticks=ticks, aspect=aspect)
     if log_scale:
         tick_labels = np.power(2, ticks)
     else:
@@ -302,12 +270,12 @@ def _figure_color_bar(exp, image, fig, axes, log_scale=True, num_ticks=8, aspect
     cb.ax.set_yticklabels(tick_labels)
 
 
-def _ax_color_bar(axes, values, width, position=0, colors=None, axis=0, label=True):
+def _ax_color_bar(ax, values, width, position=0, colors=None, axis=0, label=True):
     '''plot color bars along x or y axis
 
     Parameters
     ----------
-    axes : ``matplotlib`` axes
+    ax : ``matplotlib`` axes
         the axes to plot the color bars in.
     values : list/tuple
         the values informing the colors on the bar
@@ -356,20 +324,20 @@ def _ax_color_bar(axes, values, width, position=0, colors=None, axis=0, label=Tr
                 edgecolor="none",  # No border
                 facecolor=col[value],
                 label=value)
-            axes.add_patch(rect)
+            ax.add_patch(rect)
             if label is True:
                 rx, ry = rect.get_xy()
                 cx = rx + rect.get_width()/2.0
                 cy = ry + rect.get_height()/2.0
                 # add the text in the color bars
-                axes.annotate(value, (cx, cy), color='w', weight='bold',
+                ax.annotate(value, (cx, cy), color='w', weight='bold',
                               fontsize=7, ha='center', va='center', rotation=rotation)
         prev = i
-    # axes.legend(
+    # ax.legend(
     #     handles=[mpatches.Rectangle((0, 0), 0, 0, facecolor=col[k], label=k) for k in col],
     #     bbox_to_anchor=(0, 1.2),
     #     ncol=len(col))
-    return axes
+    return ax
 
 
 def plot(exp, sample_color_bars=None, feature_color_bars=None,
@@ -470,10 +438,10 @@ def plot(exp, sample_color_bars=None, feature_color_bars=None,
     if title is not None:
         gui_obj.figure.suptitle(title)
 
-    exp.heatmap(axes=gui_obj.axes, **kwargs)
+    exp.heatmap(ax=gui_obj.axes, **kwargs)
     barwidth = 0.3
     barspace = 0.05
-    label = color_bar_label
+
     if sample_color_bars is not None:
         sample_color_bars = _to_list(sample_color_bars)
         position = 0
@@ -481,7 +449,7 @@ def plot(exp, sample_color_bars=None, feature_color_bars=None,
             # convert to string and leave it as empty if it is None
             values = ['' if i is None else str(i) for i in exp.sample_metadata[s]]
             _ax_color_bar(
-                gui_obj.xax, values=values, width=barwidth, position=position, label=label, axis=0)
+                gui_obj.xax, values=values, width=barwidth, position=position, label=color_bar_label, axis=0)
             position += (barspace + barwidth)
     if feature_color_bars is not None:
         feature_color_bars = _to_list(feature_color_bars)
@@ -489,7 +457,7 @@ def plot(exp, sample_color_bars=None, feature_color_bars=None,
         for f in feature_color_bars:
             values = ['' if i is None else str(i) for i in exp.feature_metadata[f]]
             _ax_color_bar(
-                gui_obj.yax, values=values, width=barwidth, position=position, label=label, axis=1)
+                gui_obj.yax, values=values, width=barwidth, position=position, label=color_bar_label, axis=1)
             position += (barspace + barwidth)
     # set up the gui ready for interaction
     gui_obj()
